pyproject.toml
[project]
name = "practicerepository"
version = "0.1.0"
description = "this is just to practice pythhon and practice poetry and other libraries"
authors = [
    {name = "sai venkat",email = "saisam346@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "langchain[all] (>=0.3.27,<0.4.0)",
    "openai (>=1.107.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "transformers (>=4.56.1,<5.0.0)",
    "huggingface-hub (>=0.34.4,<0.35.0)"
]

[project.scripts]
huggingface = "huggingFaceLLMModels.console:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
